requirements.txt
git+https://github.com/iamgroot42/blosum.git
biopython
tqdm
networkx
pokemons